Made changes to Datbase Services. Added a logic check class. made chnages to Assessments, classes, and terms. Renamed a few classes as well.

+++ StudentTracker/Views/AssessmentDetails.xaml.cs
@@ -12,37 +12,46 @@ using Xamarin.Forms.Xaml;
 namespace StudentTracker.Views
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class EditAssessment : ContentPage
+    public partial class AssessmentDetails : ContentPage
     {
-        public EditAssessment()
+        
+        public AssessmentDetails()
         {
             InitializeComponent();
         }
-        public EditAssessment(Tests test)
+        public AssessmentDetails(Tests test)
         {
             InitializeComponent();
             ClassID.Text = test.ClassID.ToString();
             AssessmentID.Text = test.AssessmentID.ToString();
+            AssessmentName.Text = test.AssessmentName;
             ClassNameText.Text = test.ClassName;
             Assessment1TypePicker.SelectedItem = test.AssessmentType;
             TestStartDatePicker.Date = test.StartDate;
             TestEndDatePicker.Date = test.EndDate;
+
+           
         }
 
         async void SaveButton_Clicked(object sender, EventArgs e)
         {
-            await DatabaseService.UpdateAssessment(Int32.Parse(AssessmentID.Text), Int32.Parse(ClassID.Text), ClassNameText.Text, Assessment1TypePicker.SelectedItem.ToString(), TestStartDatePicker.Date,
+            await DatabaseService.UpdateAssessment(Int32.Parse(AssessmentID.Text), Int32.Parse(ClassID.Text), ClassNameText.Text, AssessmentName.Text, Assessment1TypePicker.SelectedItem.ToString(), TestStartDatePicker.Date,
                 TestEndDatePicker.Date);
 
             await Navigation.PopAsync();
         }
         async void DeleteButton_Clicked(object sender, EventArgs e)
         {
-            var id = int.Parse(AssessmentID.Text);
+            var confirmation = await DisplayAlert("Alert", "Are you sure you want to drop this Assessment?", "Yes", "No");
+            if(confirmation)
+            {
+                var id = int.Parse(AssessmentID.Text);
 
-            await DatabaseService.DeleteTest(id);
+                await DatabaseService.DeleteTest(id);
 
-            await Navigation.PopAsync();
+                await Navigation.PopAsync();
+            }
+           
         }
         async void CancelButton_Clicked (object sender, EventArgs e)
         {
@@ -53,5 +62,14 @@ namespace StudentTracker.Views
         {
 
         }
+
+        //private async void ShareButton_Clicked(object sender, EventArgs e)
+        //{
+        //    await Share.RequestAsync(new ShareTextRequest
+        //    {
+        //        Text = Notes.Text,
+        //        Title = "Share your notes from this course!"
+        //    });
+        //}
     }
 }
